% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_text.R
\name{encode_text}
\alias{encode_text}
\title{Encode a string into numeric positions}
\usage{
encode_text(text)
}
\arguments{
\item{text}{A character string}
}
\value{
A numeric vector
}
\description{
Encode a string into numeric positions
}
